Moved the sapi4 deprecation warning code to synthDrivers.sapi4.
