+++ source/speech/speech.py
@@ -18,12 +18,8 @@ from annotation import _AnnotationRolesT
 import controlTypes
 from controlTypes import OutputReason, TextPosition
 from controlTypes.state import State
-import globalVars
-from gui.message import MessageDialog
-import gui.contextHelp
-import queueHandler
 import tones
-from synthDriverHandler import SynthDriver, getSynth, synthChanged
+from synthDriverHandler import getSynth
 import re
 import textInfos
 import speechDictHandler
@@ -3064,43 +3060,3 @@ def clearTypedWordBuffer() -> None:
 	complete the word (such as a focus change or choosing to move the caret).
 	"""
 	_curWordChars.clear()
-
-
-def _sapi4DeprecationWarning(synth: SynthDriver, audioOutputDevice: str, isFallback: bool):
-	"""A synthChanged event handler to alert the user about the deprecation of SAPI4."""
-
-	def setShown():
-		synth._hasWarningBeenShown = True
-		synth.saveSettings()
-
-	def impl():
-		MessageDialog(
-			parent=None,
-			message=_(
-				# Translators: Message warning users that SAPI4 is deprecated.
-				"Microsoft Speech API version 4 is obsolete. "
-				"Using this speech synthesizer may pose a security risk. "
-				"This synthesizer driver will be removed in NVDA 2026.1. "
-				"You are strongly encouraged to choose a more modern speech synthesizer. "
-				"Consult the Supported Speech Synthesizers section in the user guide for suggestions.",
-			),
-			# Translators: Title of a message dialog.
-			title=_("Warning"),
-			buttons=None,
-		).addOkButton(
-			callback=setShown,
-		).addHelpButton(
-			# Translators: A button in a dialog.
-			label=_("Open user guide"),
-			callback=lambda: gui.contextHelp.showHelp("SupportedSpeechSynths"),
-		).Show()
-
-	if (not isFallback) and (synth.name == "sapi4") and (not getattr(synth, "_hasWarningBeenShown", False)):
-		# We need to queue the dialog to appear, as wx may not have been initialised the first time this is called.
-		queueHandler.queueFunction(queueHandler.eventQueue, impl)
-
-
-if not globalVars.appArgs.secure:
-	# Don't warn users about SAPI4 deprecation in secure mode.
-	# This stops the dialog appearing on secure screens and when secure mode has been forced.
-	synthChanged.register(_sapi4DeprecationWarning)
